Move mod simple into value
